Moved Concurrent Queue class into its own file
