+++ DataStructures/PBFParser.h
@@ -30,85 +30,7 @@ or see http://www.gnu.org/licenses/agpl.txt.
 #include "../typedefs.h"
 #include "HashTable.h"
 #include "ExtractorStructs.h"
-
-
-/* 
-   Concurrent Queue written by Anthony Williams: 
-   http://www.justsoftwaresolutions.co.uk/threading/implementing-a-thread-safe-queue-using-condition-variables.html  
-*/
-template<typename Data>
-class concurrent_queue
-{
-private:
-    std::queue<Data> internal_queue;
-    mutable boost::mutex queue_mutex;
-    mutable boost::mutex queue_full_mutex;
-    boost::condition_variable queue_cv;
-    boost::condition_variable queue_full_cv;
-    const size_t max_queue_size;
-
-    bool size_exceeded() const {
-        boost::mutex::scoped_lock lock(queue_mutex);
-        return internal_queue.size() >= max_queue_size;
-    }
-
-public:
-    concurrent_queue(const size_t max_size) 
-        : max_queue_size(max_size) {
-    }
-
-    void push(Data const& data)
-    {
-        if (size_exceeded()) {
-            boost::mutex::scoped_lock qf_lock(queue_full_mutex);
-            queue_full_cv.wait(qf_lock);
-        }
-
-        boost::mutex::scoped_lock lock(queue_mutex);
-        internal_queue.push(data);
-        lock.unlock();
-        queue_cv.notify_one();
-    }
-
-    bool empty() const
-    {
-        boost::mutex::scoped_lock lock(queue_mutex);
-        return internal_queue.empty();
-    }
-
-    bool try_pop(Data& popped_value)
-    {
-        boost::mutex::scoped_lock lock(queue_mutex);
-        if(internal_queue.empty())
-        {
-            return false;
-        }
-        
-        popped_value=internal_queue.front();
-        internal_queue.pop();
-        queue_full_cv.notify_one();
-        return true;
-    }
-
-    void wait_and_pop(Data& popped_value)
-    {
-        boost::mutex::scoped_lock lock(queue_mutex);
-        while(internal_queue.empty())
-        {
-            queue_cv.wait(lock);
-        }
-        
-        popped_value=internal_queue.front();
-        internal_queue.pop();
-        queue_full_cv.notify_one();
-    }
-
-    int size() const {
-        boost::mutex::scoped_lock lock(queue_mutex);
-        return static_cast<int>(internal_queue.size());
-    }
-};
-
+#include "ConcurrentQueue.h"
 
 class PBFParser : public BaseParser<_Node, _RawRestrictionContainer, _Way> {
